first steps for splitting
created a backup of scraping script, then started thinking about how a
shell script might look for split-up functions. created a “main” script
to accept command-line arguments (to be passed to sub-scripts), as well
as a “prefs” script that could host global preferences for a scraping
script if command-line arguments get too messy.